Clarify permissions for notifications and webhooks (#20741)
* Clarify permissions for notifications and webhooks

It was implied that customers could scope down notification creation access to particular products. This isn't true at the current time.

A customer with account edit permissions can create any account level notification policy, similarly a token with "Account Settings: Edit" can create any kind of notification policy.

The restrictions on creation of policies depends on the products that an account is entitled to (e.g enterprise, business, professional or free).

I also clarified that webhooks are available to everyone with a paid service, for example if you pay for Cloudflare Workers you get webhooks.

* Small edit

---------

Co-authored-by: Denise Peña <75506267+dcpena@users.noreply.github.com>